requirements.txt
Django==5.2.8
PyMySQL==1.1.2
python-dotenv==1.2.1
paho-mqtt==2.1.0
whitenoise==6.11.0
gunicorn==23.0.0
asgiref==3.11.0
sqlparse==0.5.4
tzdata==2025.2
packaging==25.0